fixed navigation and also got it so that the app stays on the correct screen even through orientation change (for supporting landscape and portrait mode
