+++ app/src/main/java/com/example/stitchcounterv3/feature/doublecounter/DoubleCounterViewModel.kt
@@ -3,19 +3,16 @@ package com.example.stitchcounterv3.feature.doublecounter
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.stitchcounterv3.domain.model.AdjustmentAmount
-import com.example.stitchcounterv3.domain.model.NavigationEvent
 import com.example.stitchcounterv3.domain.model.Project
 import com.example.stitchcounterv3.domain.model.ProjectType
 import com.example.stitchcounterv3.domain.usecase.GetProject
 import com.example.stitchcounterv3.domain.usecase.UpsertProject
+import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
@@ -35,12 +32,9 @@ open class DoubleCounterViewModel @Inject constructor(
     private val getProject: GetProject,
     private val upsertProject: UpsertProject,
 ) : ViewModel() {
+
     private val _uiState = MutableStateFlow(DoubleCounterUiState())
     open val uiState: StateFlow<DoubleCounterUiState> = _uiState.asStateFlow()
-
-    // Channel for one-time navigation events
-    private val _navigationEvents = Channel<NavigationEvent>(Channel.UNLIMITED)
-    val navigationEvents: Flow<NavigationEvent> = _navigationEvents.receiveAsFlow()
 
     fun loadProject(projectId: Int?) {
         viewModelScope.launch {
@@ -107,22 +101,9 @@ open class DoubleCounterViewModel @Inject constructor(
     }
 
     // Navigate back to main screen after saving
-    fun saveAndGoBack() {
+    fun saveAndGoBack(navigator: DestinationsNavigator) {//todo I don't think I need this at all once I have it save on back swipe
         save()
-        viewModelScope.launch {
-            _navigationEvents.send(NavigationEvent.PopBackStack)
-        }
-    }
-    
-    // Navigate to library screen
-    fun goToLibrary() {
-        viewModelScope.launch {
-            _navigationEvents.send(
-                NavigationEvent.NavigateToScreen(
-                    com.example.stitchcounterv3.feature.destinations.LibraryScreenDestination()
-                )
-            )
-        }
+        navigator.popBackStack()
     }
 }
