+++ app/src/main/java/com/example/stitchcounterv3/feature/library/LibraryScreen.kt
@@ -20,56 +20,26 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.example.stitchcounterv3.domain.model.NavigationEvent
 import com.example.stitchcounterv3.domain.model.Project
-import com.example.stitchcounterv3.feature.navigation.BottomNavGraph
+import com.example.stitchcounterv3.feature.navigation.RootNavGraph
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
-import com.ramcosta.composedestinations.navigation.navigate
-import com.ramcosta.composedestinations.navigation.popBackStack
 
-@BottomNavGraph
+@RootNavGraph
 @Destination
 @Composable
 fun LibraryScreen(
     navigator: DestinationsNavigator,
     viewModel: LibraryViewModel = hiltViewModel()
 ) {
-    // Collect navigation events and handle them
-    LaunchedEffect(viewModel) {
-        viewModel.navigationEvents.collect { event ->
-            when (event) {
-                is NavigationEvent.NavigateToScreen -> {
-                    navigator.navigate(event.destination)
-                }
-                is NavigationEvent.PopBackStack -> {
-                    navigator.popBackStack()
-                }
-                is NavigationEvent.NavigateUp -> {
-                    navigator.popBackStack()
-                }
-            }
-        }
-    }
-
     val projects by viewModel.projects.collectAsState()
     Surface(modifier = Modifier.fillMaxSize()) {
         Column(modifier = Modifier.padding(16.dp)) {
-            // Add buttons to create new projects
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                Button(onClick = { viewModel.navigateToNewSingleCounter() }) {
-                    Text("New Basic Counter")
-                }
-                Button(onClick = { viewModel.navigateToNewDoubleCounter() }) {
-                    Text("New Advanced Counter")
-                }
-            }
-            
             LazyColumn(verticalArrangement = Arrangement.spacedBy(8.dp)) {
                 items(projects) { project ->
                     ProjectRow(
                         project = project, 
-                        onOpen = { viewModel.navigateToEditProject(project) }, 
+                        onOpen = { viewModel.navigateToEditProject(project, navigator) },
                         onDelete = { viewModel.delete(project) }
                     )
                     Divider()
