+++ app/src/main/java/com/example/stitchcounterv3/feature/main/MainScreen.kt
@@ -1,5 +1,6 @@
 package com.example.stitchcounterv3.feature.main
 
+
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -9,7 +10,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
@@ -17,37 +17,19 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.stitchcounterv3.R
-import com.example.stitchcounterv3.domain.model.NavigationEvent
-import com.example.stitchcounterv3.feature.navigation.BottomNavGraph
+import com.example.stitchcounterv3.feature.destinations.DoubleCounterScreenDestination
+import com.example.stitchcounterv3.feature.destinations.SingleCounterScreenDestination
+import com.example.stitchcounterv3.feature.navigation.RootNavGraph
 import com.example.stitchcounterv3.ui.theme.StitchCounterV3Theme
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 
-
-@BottomNavGraph(start = true)
+@RootNavGraph(start = true)
 @Destination
 @Composable
 fun MainScreen(
-    viewModel: MainViewModel = hiltViewModel(),
     navigator: DestinationsNavigator
 ) {
-    // Collect navigation events and handle them
-    LaunchedEffect(viewModel) {
-        viewModel.navigationEvents.collect { event ->
-            when (event) {
-                is NavigationEvent.NavigateToScreen -> {
-                    navigator.navigate(event.destination)
-                }
-                is NavigationEvent.PopBackStack -> {
-                    navigator.popBackStack()
-                }
-                is NavigationEvent.NavigateUp -> {
-                    navigator.popBackStack()
-                }
-            }
-        }
-    }
-
     Surface(modifier = Modifier.fillMaxSize()) {
         Column(
             modifier = Modifier
@@ -57,10 +39,14 @@ fun MainScreen(
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
             Text(stringResource(id = R.string.app_name), style = MaterialTheme.typography.headlineMedium)
-            Button(onClick = { viewModel.navigateToSingleCounter() }) {
+            Button(onClick = {
+                navigator.navigate(SingleCounterScreenDestination())
+            }) {
                 Text("New Single Tracker")
             }
-            Button(onClick = { viewModel.navigateToDoubleCounter() }) {
+            Button(onClick = {
+                navigator.navigate(DoubleCounterScreenDestination())
+            }) {
                 Text("New Double Tracker")
             }
         }
@@ -91,4 +77,3 @@ private fun MainScreenPreview() {
         }
     }
 }
-
