+++ app/src/main/java/com/example/stitchcounterv3/feature/navigation/BottomNavTab.kt
@@ -5,6 +5,7 @@ import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.List
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.ui.graphics.vector.ImageVector
+import com.ramcosta.composedestinations.annotation.Destination
 
 enum class BottomNavTab(
     val title: String,
