+++ app/src/main/java/com/example/stitchcounterv3/feature/navigation/BottomNavigationLayout.kt
@@ -1,0 +1,97 @@
+package com.example.stitchcounterv3.feature.navigation
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.slideInVertically
+import androidx.compose.animation.slideOutVertically
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.NavigationBarItemColors
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.navigation.NavHostController
+import com.example.stitchcounterv3.feature.NavGraphs
+import com.example.stitchcounterv3.feature.destinations.LibraryScreenDestination
+import com.example.stitchcounterv3.feature.destinations.MainScreenDestination
+import com.example.stitchcounterv3.feature.destinations.SettingsScreenDestination
+import com.ramcosta.composedestinations.DestinationsNavHost
+import com.ramcosta.composedestinations.navigation.navigate
+
+
+@Composable
+fun BottomNavigationLayout(
+    selectedTab: BottomNavTab,
+    onTabSelected: (BottomNavTab) -> Unit,
+    navController: NavHostController
+) {
+    AnimatedVisibility(
+        visible = true,
+        enter = slideInVertically(
+            initialOffsetY = { fullHeight -> fullHeight }
+        ),
+        exit = slideOutVertically(
+            targetOffsetY = { fullHeight -> fullHeight }
+        )
+    ) {
+        BottomNavigationBar(
+            selectedTab = selectedTab,
+            navController = navController,
+            onTabSelected = onTabSelected
+        )
+    }
+}
+
+@Composable
+private fun BottomNavigationBar(
+    selectedTab: BottomNavTab,
+    navController: NavHostController,
+    onTabSelected: (BottomNavTab) -> Unit
+) {
+    NavigationBar {
+        BottomNavTab.entries.forEach { tab ->
+            NavigationBarItem(
+                selected = selectedTab == tab,
+                onClick = {
+                    onTabSelected(tab)
+                    val destination = when(tab) {//todo repeatedCode
+                        BottomNavTab.HOME -> MainScreenDestination
+                        BottomNavTab.LIBRARY -> LibraryScreenDestination
+                        BottomNavTab.SETTINGS -> SettingsScreenDestination
+                    }
+                    navController.navigate(destination) {
+                        // Avoid building up a huge stack
+                        popUpTo(navController.graph.startDestinationId) {
+                            saveState = true
+                        }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
+                },
+                icon = {
+                    Icon(
+                        imageVector = tab.icon,
+                        contentDescription = tab.title
+                    )
+                },
+                label = {
+                    Text(tab.title)
+                },
+                colors = NavigationBarItemColors(
+                    selectedIconColor = MaterialTheme.colorScheme.secondary,
+                    selectedTextColor = MaterialTheme.colorScheme.secondary,
+                    unselectedIconColor = Color.Gray,
+                    unselectedTextColor = Color.Gray,
+                    selectedIndicatorColor = MaterialTheme.colorScheme.primary.copy(alpha = 0.2f),
+                    disabledIconColor = Color.Gray,
+                    disabledTextColor = Color.Gray
+                )
+            )
+        }
+    }
+}
