+++ app/src/main/java/com/example/stitchcounterv3/feature/navigation/NavGraphs.kt
@@ -5,6 +5,3 @@ import com.ramcosta.composedestinations.annotation.NavGraph
 @NavGraph
 annotation class RootNavGraph(val start: Boolean = false)
 
-@NavGraph
-annotation class BottomNavGraph(val start: Boolean = false)
-
