+++ app/src/main/java/com/example/stitchcounterv3/feature/navigation/RootNavigationViewModel.kt
@@ -1,7 +1,6 @@
 package com.example.stitchcounterv3.feature.navigation
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -9,7 +8,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
 
 @HiltViewModel
-class BottomNavigationViewModel @Inject constructor() : ViewModel() {
+class RootNavigationViewModel @Inject constructor() : ViewModel() {
     
     private val _selectedTab = MutableStateFlow(BottomNavTab.HOME)
     val selectedTab: StateFlow<BottomNavTab> = _selectedTab.asStateFlow()
